More Unboxing

+++ src/net/sourceforge/kolmafia/session/TavernManager.java
@@ -385,7 +385,7 @@ public class TavernManager {
     Integer searchIndex = searchList.remove(0);
 
     // That's the square we will visit.
-    return searchIndex.intValue();
+    return searchIndex;
   }
 
   public static void handleTavernChange(String responseText) {
